+++ src/main/java/net/silentchaos512/gear/item/gear/CoreArmor.java
@@ -32,7 +32,7 @@ import javax.annotation.Nullable;
 import java.util.List;
 import java.util.UUID;
 
-public class CoreArmor extends ArmorItem implements ICoreArmor {
+public class CoreArmor extends DyeableArmorItem implements ICoreArmor {
     // Just copied from ArmorItem, access transformers are too flaky
     private static final UUID[] ARMOR_MODIFIERS = {UUID.fromString("845DB27C-C624-495F-8C9F-6020A9A58B6B"), UUID.fromString("D8499B04-0E66-4726-AB29-64469D734E0D"), UUID.fromString("9F3D476D-C118-4544-8365-64846904B48E"), UUID.fromString("2AD3F246-FEE1-4E67-B886-69FD380BB150")};
     // sum = 1, starts with boots
@@ -170,13 +170,6 @@ public class CoreArmor extends ArmorItem implements ICoreArmor {
                 + ".png";
     }
 
-    // FIXME
-    /*
-    @Override
-    public boolean hasOverlay(ItemStack stack) {
-        return true;
-    }
-
     @Override
     public boolean hasColor(ItemStack stack) {
         return true;
@@ -193,7 +186,6 @@ public class CoreArmor extends ArmorItem implements ICoreArmor {
 
     @Override
     public void setColor(ItemStack stack, int color) {}
-    */
 
     @Override
     public boolean hasEffect(ItemStack stack) {